+++ app/Http/Controllers/ProductsController.php
@@ -2,14 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use App\Exports\ProductExport;
+use App\Imports\ProductImport;
 use App\Models\Product;
 use App\Models\Supplier;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use App\Services\Product\StoreProductService;
 use App\Http\Requests\Product\StoreProductRequest;
 use App\Http\Requests\Product\UpdateProductRequest;
 use App\Services\Product\UpdateProductService;
+use Illuminate\Http\Request;
+use Maatwebsite\Excel\Facades\Excel;
 
 class ProductsController extends Controller
 {
@@ -26,18 +29,28 @@ class ProductsController extends Controller
 
     public function index()
     {
+        if (request()->user()->cannot('index', Product::class)) {
+            abort(403);
+        }
         $products = Product::with('supplier')->get();
         return view('products.index', \compact('products'));
     }
 
     public function create()
     {
+        if (request()->user()->cannot('create', Product::class)) {
+            abort(403);
+        }
         $suppliers = Supplier::select('id', 'trade_name')->get();
         return view('products.create', \compact('suppliers'));
     }
 
     public function edit($id)
     {
+        $product = new Product();
+        if (request()->user()->cannot('update', $product)) {
+            abort(403);
+        }
         $row = Product::findOrFail($id);
         $suppliers = Supplier::select('trade_name', 'name', 'id')->get();
         return response()->json([
@@ -48,6 +61,9 @@ class ProductsController extends Controller
     public function store(StoreProductRequest $request)
     {
         try {
+            if (request()->user()->cannot('create', Product::class)) {
+                abort(403);
+            }
             $data = $request->validated();
 
             $this->storeProductService->store($data);
@@ -67,6 +83,10 @@ class ProductsController extends Controller
     public function update(UpdateProductRequest $request, $id)
     {
         try {
+            $product = new Product();
+            if (request()->user()->cannot('update', $product)) {
+                abort(403);
+            }
             $this->updateProductService->update($request->validated(), $id);
             return response()->json([
                 'message' => 'تم تعديل المادة بنجاح.',
@@ -84,6 +104,10 @@ class ProductsController extends Controller
     public function destroy($id)
     {
         try {
+            $product = new Product();
+            if (request()->user()->cannot('delete', $product)) {
+                abort(403);
+            }
             Product::where('id', $id)->delete();
             return response()->json([
                 'message' => 'تم حذف المادة بنجاح',
@@ -101,6 +125,9 @@ class ProductsController extends Controller
     public function MultiDelete(Request $request)
     {
         try {
+            if (request()->user()->cannot('MultiDelete', Product::class)) {
+                abort(403);
+            }
             Product::whereIn('id', (array) $request['ids'])->delete();
             return response()->json([
                 'message' => 'تم حذف المواد بنجاح',
